requirements.txt
fastapi
uvicorn
pydantic>=2
pyyaml
httpx
cryptography
python-dotenv
bcrypt

numpy
